// apps/web/src/lib/api/generated-client/models/message-create-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI No description provided (generated by Openapi Generator
 * https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator
 * (https://openapi-generator.tech). https://openapi-generator.tech Do not edit
 * the class manually.
 */

import { mapValues } from '../runtime';
import type { AgentConfig } from './agent-config';
import {
  AgentConfigFromJSON,
  AgentConfigFromJSONTyped,
  AgentConfigToJSON,
  AgentConfigToJSONTyped
} from './agent-config';
import type { MessageCreateInputContentInner } from './message-create-input-content-inner';
import {
  MessageCreateInputContentInnerFromJSON,
  MessageCreateInputContentInnerFromJSONTyped,
  MessageCreateInputContentInnerToJSON,
  MessageCreateInputContentInnerToJSONTyped
} from './message-create-input-content-inner';

/**
 * @export
 * @interface MessageCreateInput
 */
export interface MessageCreateInput {
  /**
   * @memberof MessageCreateInput
   * @type {MessageCreateInputContentInner[]}
   */
  content: Array<MessageCreateInputContentInner>;
  /**
   * The configuration for the agent that generated the message. Requires at
   * least a `agent_class` key.
   *
   * @memberof MessageCreateInput
   * @type {AgentConfig}
   */
  agent_config: AgentConfig;
}

/** Check if a given object implements the MessageCreateInput interface. */
export function instanceOfMessageCreateInput(
  value: object
): value is MessageCreateInput {
  if (!('content' in value) || value['content'] === undefined) return false;
  if (!('agent_config' in value) || value['agent_config'] === undefined)
    return false;
  return true;
}

export function MessageCreateInputFromJSON(json: any): MessageCreateInput {
  return MessageCreateInputFromJSONTyped(json, false);
}

export function MessageCreateInputFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): MessageCreateInput {
  if (json == null) {
    return json;
  }
  return {
    content: (json['content'] as Array<any>).map(
      MessageCreateInputContentInnerFromJSON
    ),
    agent_config: AgentConfigFromJSON(json['agent_config'])
  };
}

export function MessageCreateInputToJSON(json: any): MessageCreateInput {
  return MessageCreateInputToJSONTyped(json, false);
}

export function MessageCreateInputToJSONTyped(
  value?: MessageCreateInput | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    content: (value['content'] as Array<any>).map(
      MessageCreateInputContentInnerToJSON
    ),
    agent_config: AgentConfigToJSON(value['agent_config'])
  };
}